testsuite: fix check for qmanager load failure
Problem: test expects flux module load to fail when a hello
callback fails, but hello callbacks now occur asynchronously.

Change the test to look for telltale failures in the broker log.

Drop the check for 'flux module remove qmanager' failure in the
following test, since that could now be racy and is somewhat
redundant with the broker log check.