resource: add support for node-local storage to hwloc reader
Problem: versions 2+ of hwloc emits OS device objects that represent
node-local block devices, but the resource hwloc reader does not pick
those up when they exist.

Solution: add a new check for `HWLOC_OBJ_OSDEV_BLOCK` in the
`walk_hwloc` method and parse out the size from the hwloc object
properties.  The name of the resource in the graph is `storage`.