t: unify storage logical index across hwloc V1 and V2
Problem: the logical indexing for the `OSDev`s is not consistent across
major hwloc versions.  There seems to be a different sorting that is
applied.  The end result is that the hwloc-based tests fail their string
literal comparison due to differing logical IDs:

```
--- /usr/src/t/data/resource/expected/node_local_storage/catalyst.R.out
2020-08-03 23:59:05.000000000 +0000
+++ catalyst.R.out      2020-08-04 00:06:45.000000000 +0000
@@ -10,10 +10,10 @@
       ------------core21[1:x]
       ------------core22[1:x]
       ------------core23[1:x]
-      ------------storage8[195:x]
-      ------------storage9[195:x]
       ------------storage10[195:x]
       ------------storage11[195:x]
+      ------------storage12[195:x]
+      ------------storage13[195:x]
       ---------socket1[1:x]
       ------catalyst321[1:s]
       ---cluster0[1:s]
not ok 3 - match allocate with catalyst hwloc and catalyst jobspec
--- /usr/src/t/data/resource/expected/node_local_storage/corona.R.out
2020-08-03 23:59:05.000000000 +0000
+++ corona.R.out        2020-08-04 00:11:09.000000000 +0000
@@ -4,7 +4,7 @@
       ------------core3[1:x]
       ------------core4[1:x]
       ------------core5[1:x]
-      ------------storage0[1562:x]
+      ------------storage1[1562:x]
       ------------core6[1:x]
       ------------core7[1:x]
       ------------core8[1:x]
```

Solution: remove from t/data/hwloc-data/node_local_storage hwloc XML
files the other `OSDev`s that appear logically before the storage
`OSDev`.  In this way, the logical IDs of the storage resources will be
the same across hwloc versions for this test, without affecting the
matching within fluxion.