testsuite: flux-ion-resource.py: use JobID class for jobids
Problem: t/scripts/flux-ion-resource.py subcommands that take JOBID
arguments restrict the jobid type to `int`, but there are many other
possible jobid representations that could be conventent to use with
the utility.

Change the `type` of jobid args in flux-ion-resource.py from `int` to
the Flux `JobID` type. This allows a jobid in any supported encoding
to be ingested by the tool