resource: add `LinuxDeviceID` property to `storage` resource vertex
Problem: The device major and minor IDs are required in order to
throttle and restrict the view to block storage devices.  Hwloc collects
and reports this information, but it is not stored in the fluxion
resource graph or in the output JGF.

Solution: when a block storage device is encountered, get the
`LinuxDeviceID` (e.g. `8:16`) from hwloc and store it in the vertex's
`property` map.  This `property` map is already emitted by the JGF
writer and read by the JGF reader.

Further reading:
- Linux cgroup device whitelist controller:
https://www.kernel.org/doc/html/latest/admin-guide/cgroup-v1/devices.html
- Linux cgroup block IO controller:
https://www.kernel.org/doc/html/latest/admin-guide/cgroup-v1/blkio-controller.html