traverser: Better handling if cnt_slot() returns bad info
Problem: If `cnt_slot()` incorrectly counts the number
of available slots, its downstream code can lead
to a buffer overflow within `dfu_slot`. Because
`cnt_slot()` depends on the user-provided jobspec
input, this can be exploited or at least lead to a
less robust or undefined behavior.

Stop iterating `dfu_slot` when you no longer have
available slot and immediately return "no match".
Because adding this iteration was awkward with the current
scoring API interface, rewrite the iteration API
as part of this as well.