resource: propagate labels in jobspec through to R
Problem: it is hard to put together exactly which resources in jobspec
correspond with which resources in R.  The connection almost amounts to
reperforming the full scheduling match again.  For components like job
shell plugins, this imposes unnecessary complexity.

Solution: allow the user to provide a `label` for resources within
jobspec, and then propagate that label through to R.  This allows for
the easy connection of resources in jobspec with resources in R.

In order to propagate the label in jobspec through to R within Fluxion,
we need to add a new data structure to each resource vertex, called
`ephemeral`.  This data structure is intended to store data that is only
valid for the current match being performed. When a resource in the
graph matches a resource in jobspec (and the jobspec resource has a
label), we insert the jobspec resource label into the graph vertex's
`ephemeral` data structure (along with the current epoch, which is used
for memory management, described below).  The match policy is then
applied to select a subset of the matching resources for the final
allocation.  Finally, any resources selected by the match policy are
`emit`ted, which includes the emission of the label stored in the
`ephemeral` data structure.

Note on memory management: The `ephemeral` data structures for selected
vertices are cleared after the vertices are emitted.  For resources that
were matched but not selected, the label still resides in the
`ephemeral` data structure.  The previously mentioned epoch is used to
ensure that no stale data are ever returned later on.  If a resource
vertex has been selected and has stale data, the stale data are cleared
only after the vertex has been selected and emitted.  This lazy clearing
of stale data is done to avoid traversing the entire graph clearing
`ephemeral` members after every resource query.