testsuite: use new test functions
Change tests that unload sched-simple to set FLUX_SCHED_MODULE=none.

Instead of reloading hwloc with test XML, use the load_test_resources()
function to do that.  The function ensures that the core resource module
and fluxion are in sync, and allows the mechansim to be changed in one
place as the core resource module evolves.

Ensure that cleanup_active_jobs() is called before scheduler is
unloaded in any tests that run jobs, to ensure that there is no hang
in the core cleanup scripts (with no scheduler).