testsuite: use load_test_resources() in tests
Fix a few tests that are still using 'flux hwloc reload' directly
to change the configured resources to use the load_test_resources()
shell function.

In t6002-graph-hwloc.t, this fixed a test synchronization issue
introduced by upstream changes to the resource module, in which
not all brokers were yet marked up when the fluxion resource
module is loaded in the test.