testsuite: change method of setting test resources
Problem: verification of fake test resources fail with
latest changes to the core resource module.

The core resource module now checks that probed resources
match configured resources.  This check must be disabled
by loading the resource module with the 'noverify' option
in test.

Also, it is no longer necessary to post a 'hwloc-discover-finish'
event in the resource.eventlog when setting up fake resources.

Also, the fake resource object (still in by_rank format at this time)
needs to be located in the KVS at resource.R.