writer: add an additonal arg to emit_json method
Problem: to comply with the latest RFC20 change,
host names should no longer be emitted by the rlite
writer. However, the rlite writer is one that
sees all the host names and need a way to pass
them optionally to the caller. With the current
"emit_json" interface with one output method
arguement, there isn't really an easy way
to pass such optional info.

Extend the emit_json method in the base writer
abstract class and have all of its derived classes
to override it.

The rlite writer class will be the first class
that will make use of this additional "aux"
method argument. But this will likley reduce potential
future interface churn for other writer classes
as well.