writer: id compression takes vector instead of set
Problem: This ID compress function will be used
more generally within the writer classes. But
taking std::set as the input container makes it
a bit less general. There will be cases where
we need to compress IDs in the exact order
in which they are added to the container.

Use std::vector as the input container.
This means the users of this method must sort
it before using the compression method
if the ID compression should be done
on the sorted IDs, which is the original
behavior. As such, this is a trade-off between
performance and usability.