testsuite: convert 04-brokers test XML input
Problem: `flux resource reload -x xmldir` loads XML int the flux-core
resource module unmodified, and provides this XML to users of the
resource.get-xml RPC. This is different from the old `flux hwloc
reload` method, which loaded XML into a hwloc_topology_t and re-emitted
it, in the process potentially converting or modifying the XML in
some way. The t6002-graph-hwloc.t test appears to be sensitive to
the fact that this conversion is no longer occuring automatically,
and so the test fails with extra vertices detected:

  expecting success:
      flux ion-resource stat > stat.out &&
      verify stat.out
  2c2
  <   "[0-3]": 71
  ---
  >   "[0-3]": 37

Fix the issue by converting the input XML using the hwloc-convert
utility from flux-core (t/hwloc/hwloc-convert). Which reads in
the XML with hwloc 2.x, and then emits new XML using the flag
HWLOC_TOPOLOGY_EXPORT_XML_FLAG_V1. The XML will then be equivalent
to the expected "pre-converted" input.