utilities: add attach to resource-query
Add an attach command to the resource-query tool.
Attach takes a fully-qualified JGF subgraph as
an argument. After unpacking the subgraph,
the resource graph must be reinitialized
so that the pruning filters are properly
initialized with the new resources.
Since reinitialization creates new
subtree plans, allocated and reserved
jobs may be affected. The command
exits with an error if there are
currently allocated or reserved jobids.

The attach command also enables testing of
the unpack_at() functionality.