reader: add unpack_at functionality
Add the ability to add a JGF subgraph into
a resource graph by implementing the `unpack_at()`
function. The function takes a fully-qualified
JGF subgraph and an attachment point as inputs.
The JGF is said to be fully qualified in that
every vertex has a path to the root via
intermediate vertices. The root and intermediate
vertices are also in the JGF subgraph. Currently,
the vertex attachment point is not used by
`unpack_at()`, but could be used to speed
up the unpacking.

Remove the check for empty `is_roots` map
in `update_tgt_edge()`. The check causes a
spurious error when encountering an edge
into a root before the edge out of a root
in a JGF with bidirectional edges. We
cannot assume an edge iteration ordering,
so the check must be removed for robustness.

Refactor a commonly used sequence of code
into a function (`update_vmap`) for readibility.