reader: add resource namespace remapper support
Problem: The rank and resource Id namespace of a
nested flux instance is different from that of its
parent. Thus, the resource reader of a nested
instance may need to remap the rank and/or certain
resource Ids as it populates its resource graph
store from the resources emitted from its parent.

Add a header and C++ source file that introduces
the resource_namespace_remapper_t class. It is
designed to add and query resource Id
or rank remapping information efficiently
both in terms of performance and memory overheads.

Using the add() method, users should first pass
the resource type (e.g., "core", "gpu") or "rank" as
the "name_type" argument along with the Id remapping
info (ref_id is remapped to remapped_id) for each
unique range of flux ranks (e.g., 0-3).

resource_namespace_remapper_t keeps each
remapping data using distinct_range_t as key to a
std::map member object. IOW, we treat each distinct
rank range which has the same remapping info as
an equivalent set and keep only one copy
of remapping info.

Furthermore, the distinct_range_t class implements
"operator<()" so that this mapping info can be kept
in std::map in ascending rank-range order,
which then allows data querying to be performed
in O(logN) time.

However, the method for querying is the query() method,
whose semantics is per individual rank based.
To support retrieval of remapping information
for a individual rank, we implement "operator<()"
in such a way that we can find the corresponding
record when the rank intersects a rank-range key.