testsuite: add tests for unpack_at
Add tests for `unpack_at()`. The tests
add a JGF subgraph with default
and then randomized vertex and edge
orderings. The tests check for
error conditions such as attempting
to add another graph root and
attaching a JGF subgraph with
outstanding job allocations or
reservations, as well as basic
checks for correctly formed
command arguments.