traverser: fix segfault for bi-edge traversal
Fix a non-terminating traversal cycle for
depth-first traversal of a graph with
bidirectional edges. Upon reaching a leaf,
the depth-first traverser would traverse
the out edge back to the ancestor, which
in turn traverses back to the leaf in
a non-terminating cycle until producing
a segfault. Inserting a gray coloration
breaks the cycle.