reader: support for remap-aware vertex unpack
Add support to remap exectution target and
compute core Ids when namespace_remapper is
available.

The scheduling section of RV1 (in JGF format)
as received from core's resource.acquire
interface is in the parent instance's
namespace (unless the instance is at the
top level). Thus, its Id namespace must
be remapped to that of the current instance.
Because how Ids are generated, only
the execution target and core Ids
currently need to be remapped.
(e.g., GPU Ids are always expressed
in the global namespace).

Note that we rewrite the path information as
it contains the resource's basename and Id.