resource: delay creating reader object
Problem: the resource reader object is currently created
too early such that it precludes us from picking the
right reader type based on the response from core's
resource.acquire interface, which happens much
later in time.

Refactor the creation of resource reader object
in a function and delay the calling of this function
until we get to know what resource reader type we need.