resource: use rv1+JGF to populate graph data store
Use the scheduling key (JGF) of rv1 sent from
core's resource.acquire interface to populate
the graph data store.

Continue to use the hwloc resource reader if
the scheduling key is non-existent in the rv1
object.

Remap the the resource and execution target Ids
in the JGF object to the current Flux instance's
namespace because the embedded JGF is not remapped
(IOW, it is in the parent's namespace).

Because the execution key of the rv1 object
does not contain the requisite information for us
to be able to do this parent-child namespace
remapping, we fetch and use the job rv1 object
from the parent's namespace.