reader: refactor get_low_high () into distinct_range_t
Problem: It is awkward to pull in resource_namespace_remapper_t
just to use get_low_high ().

Make this method as a static method of distinct_range_t
so that only distinct_range_t can be used without
having to pull in resource_namespace_remapper_t.

Also modify the add_exec_target_range in
the resource_namespace_remapper_t class so that
a distinct_range_t object is passed as the
second parameter of this method. This will
help simplify the user code of this class
(i.e., sched-fluxion-resource) a bit.