testsuite: refactor remap_rv1_resource_type
Move remap_rv1_resource_type into the common test
script (sched-sharness.sh) as this will be used
in additional tests in the future.

Change this shell function such that it can take
in rank remapping offset for each and all 4 ranks
and apply it using conditionals in the jq filter.