testsuite: flux-ion-R.py to generate JGF-embedded RV1
Add flux-ion-R.py to help test Fluxion with the inital
RV1 object that contains the scheduling section.

Take in an RV1 JSON string via stdin and use
jsongraph to generate and add the corresponding
JGF into the scheduling key of the input RV1.

Print the resulting RV1 (with the embedded JGF)
out to stdout.

Example usage:
% flux R encode --hosts=sierra[3682,3179,3683,3178] \
--cores=0-43 --gpu=0-3 | flux ion-R > sierra.R.test