traverser: enable resource graph reinitialization
Add the ability to reinitialize the resource
graph. After attaching a JGF subgraph to the
original resource graph, the pruning filters
must be reinitialized with the new resources.

Add `reset_color()` to the DFU Traverser
Implementation Public API. This function
allows the DFU traverser to reset the
graph color before reinitializing the
resource graph. Without this function,
`dfu_impl_t::prime_pruning_filter()`
stops at the root vertex for a resource
graph that has been initialized.

Note that the reinitialization currently
traverses the entire resource graph. In the
future we will need to add an optimization
to only traverse and initialize the added
subgraph.