t7000-shell-datastagint.t: fix path to flux-shell
Problem: The datastaging test sets the path to flux-shell based
on pkg-config output, but the PKG_CONFIG_PATH may not necessarily
contain pkg-config files for the version of flux-core under which
the test is running. This is because PKG_CONFIG_PATH is not amended
when running under flux(1). This is especially true when building
flux-sched against a side install version of flux-core. Then the path
to the flux-shell will most likely pick up the system installed flux
instead of the side-install version.

Use the test script flux-builtin-conf.py to get the builtin config for
`shell_path` instead of using pkg-config. Since this pulls directly
from libflux, it should exactly match the shell path of the currently
running Flux, not the system version.

Ensure that the path is discovered *after* running under Flux, so
that we're definitely using the intendted libflux.