queue: add schedulability support to base queue class
Introduce schedulability support in prepration for
the upcoming delay scheduling optimization.

The delay scheduling feature moves the invocation of
the schedule loop into prepare/check watcher
callbacks. In other words, the running of the schedule
loop will become asynchronous with respect
to job and resource events that must cause this
loop to be executed.

Build schedulability state tracking directly into the
base queue class so that whether to run the
schedule loop or not can asynchronously be checked.