traverser: add by_subplan pruning for headless jobspecs
Problem: for the class of jobspecs whose first resource
type in the resource section is "slot", aggregate
based by_subplan prunning is not performed. This leads
to poor performance for this jobspec class.

Instead of giving up pruning when visiting resource
is not matched with the currently processing
resource of the jobspec, make a special case
for resource.type == slot.

The prune () method is only called for ether
true match or "slot" match, this allows us to
use user_data at the "slot" resource level
to do by_subplan check. The only thing to
be careful is not to do by_excl check in this case.