traverser: update out-edges metadata per allocation
Update the newly added out-edges metadata
with all of the edges used in a resource
allocation.
Perform this every time an allocation is created.

Use the currently available resource amounts
of the pruning-filter resource type
as the weight of each outgoing edge.
Specifically, we use the last resource type
in the pruning filter multi-planner.

E.g., if no additional pruning filter
type is not used, this will be "core"; otherwise
the last resource type in the additional
types added to the pruning filter is used.

The rationale is: users would only add an additonal
filter type (e.g., gpu) as this is likely more
scarce resource type than "core"; and for
our first match policy, we will want
to visit vertices with the most amounts
of the scarcest resource type first.