query: add multi-match command
Add the multi-match command into resource-query
to be consistent with the corresponding
RPC recently added to sched-fluxion-resource.

Allocate or reserve for multiple jobspecs.
Provide the same subcommands as that of the singletone
match command: i.e., allocate,
allocate_with_satisfiability, and allocate_orelse_reserve.

resource-query> multi-match allocate jobspec1 jobspec2 ...

The returning semantics is consistent with that
of the RPC version. When it finds no matching resource
for a jobspec, it returns without having
to process later jobspecs.