github: ensure annotated tag is available for release
Problem: When the GitHub action runs to deploy a tagged release, it
doesn't use the correct version string. This causes the tarball to
fail to upload, since the tarball version doesn't match the expected
version. This is because the default checkout@v2 action creates
lightweight tags instead of annotated.

See also https://github.com/actions/checkout/issues/290

If the GitHub action is creating a release, fetch the corresponding
annotated tag directly using the suggestion in the referenced issue
above, allowing `git describe` to return the correct tagged version.