qmanager: add provisional cancel support
Problem: We currently don't handle sched.cancel
correctly under the new async execution support.
The semantics used to be that when a sched.cancel
is received by qmanager, the pending job state
was deterministic.
It is no longer true because there is a maybe-pending
state: a request has sent out
to sched-fluxion-resource.multi-match and
the corresponding stream RPC has not been received.
Because mishandling this, a race condition
can arise leading to a crash.

Detect "maybe-pending" state when the
sched.cancel callback is invoked.

Add this request to the newly introduced
provisional cancel queue.

When the current schedule loop invocation
completes, handle each and every cancel
request queued up by moving the target
pending jobs to the newly added canceled queue.

In post_sched_loop, each and every
canceled jobs' alloc request is denied in order.