build: set automake foreign mode
Problem: COPYING is missing, but autmake assumes it is required
because it assumes every project follows the GNU project template.

autogen.sh creates COPYING because it calls automake with the
--create-missing option, but if automake is invoked by make rules,
it fails.

Set automake foreign mode, indicating that this project does not
follow the GNU template that requires COPYING and other files.