resource: add a method to calculate effective max request
Problem: Depending on the operator and operand of the
count key of a Jobspec, the value specified
in the max key may not be satisfiable. For example,
if min=1/max=5 and operator="*"/operand=2, the
effective satisfiable max is 4, not 5. There are cases
where calculating an effective max can come in handy.

Add the calc_effective_max() method to
the matcher_util_api_t class for future use.