jobspec: make libjobspec comply with RFC14 for optional keys
Problem: libjobspec within Fluxion currently requires all
four keys for a "moldable" jobspec to be present under
the resources.count: min, max, operator, and operand.
This does not comply with RFC14, as it specifies:

"The default value for max SHALL be infinite, therefore
a count which specifies only the min key SHALL be considered
a request for at least that number of a resource, and
the scheduler SHALL generate the R that contains the maximum
number of the resource that is available and subject to the
operator and operand. By contrast, if a fixed count is given
to the count key, the scheduler SHALL match any resource that
contains at least count of the resource, but its R SHALL
contain exactly count of the resource (potentially leaving
excess resources unutilized).

Handle the optional keys (max, operator, operand) appropriately
and adjust the relevant test cases.