docker: set FLUX_ENABLE_VALGRIND_TEST during checks
Problem: When running docker-run-checks.sh, the valgrind test could
be skipped (unless tests run with debug=t) since that test is now
optional, but the valgrind test is known to have correct suppressions
in the purpose-built docker container so the test should not be
skipped.

Set FLUX_ENABLE_VALGRIND_TEST in checks_run.sh to ensure the valgrind
test runs when using docker-run-checks.sh.