policy: add node-locality-aware match policy
Make use of the new multilevel_id_t and add a
node-level containst by adding a score factor
for the node resource type to this class.

Net effect: all of the node-local resource will
have this node-level factor applied.

Combining this with high_id_first, which
is called the "hinode" policy:
Within node-level locality constraint, the resource vertex
with higher ID is preferred among its kind
(e.g., core1 is preferred over core0 but those in node1
       is selected first before those in node0 is selected)
If a node-level locality contraint is already specified in
the jobspec, this policy is identical to high: e.g.,
    node[2]->slot[2]->core[2].
Only when node-level locality contraint is absent, the
match behavior is deviated from high: e.g.,
    slot[2]->core[2].

Combining this with low_id_first, which
is called the "lonode" policy:
Within node-level locality constraint, the resource vertex
with lower ID is preferred among its kind
(e.g., core0 is preferred over core1 but those in node0
is selected first before those in node1 is selected)
If a node-level locality contraint is already specified in
the jobspec, this policy is identical to low: e.g.,
    node[2]->slot[2]->core[2].
Only when node-level locality contraint is absent, the
match behavior is deviated from low: e.g.,
    slot[2]->core[2].