traverser: change an overloaded method name
Problem: The traverser class overloads resolve
methods, one for resolving a visiting vertex
and the other for resolving an entire graph
at the end of each traversal. Because
the operational scope between two methods
is quite different, this overloading can be
confusing for future modification.

Change the method name of the graph-scope
resolution to "resolve_graph."