mergify: update configuration to replace deprecated keys
This commit message is adapted from the mergify update
commits in flux-core.

Problem 1: Mergify.io has deprecated the strict merge mode action:

 https://blog.mergify.io/strict-mode-deprecation/

Replace the strict mode configuration with a default queue with a
rebase action which should be equivalent.

Problem 2: The current mergify config specifies that PRs are merged
using a rebase method and not merge. This would result in the PR branch
being rebased onto the main branch instead of creating a merge commit,
which we use to generate release notes. Luckily, flux-sched branch
protection rules prevent the rebase from being completed by mergify.

Change the queue action for PRs to be merge, with an update_method of
rebase, which should be similar to the previous 'strict' mode flux-sched
was using earlier.