optmgr: Add jsonify support to optmgr
Problem: It is currently difficult to examine the module option
parameter values passed to qmanager. This will become worse as we plan
to add TOML configuration support to sched-fluxion-resource soon as
well.

Add the jsonify method of optmgr composer class so that each of the
module-specific option management classes on which a composer class is
templated must implement ways to export its managing option parameter
values as a JSON string.