qmanager: add subtable support for sched-fluxion-qmanager
Problem: Currently the "queue-params" and "policy-params" keys in the
sched-fluxion-qmanager TOML config file support takes a comma-separated
list of parameters. It was noted from a review, it would be much cleaner
and more extensible to hoist each of "queue-params" and "policy-params"
as a sub-table and put the comma-separated parameters as formal key
value pairs.

Modify qmananger.cpp to support sub-tables.

Modify the manual for qmanager config as well as test cases.