testsuite: add match allocate test with AMD RSMI GPUs
Problem: No tests in the Fluxion testsuite ensure that AMD RSMI GPUs
are detected by the hwloc reader and available for match/allocaate.

Add a HWLOC XML from a system with 8 AMD GPUs detected by RSMI and
use this xml for a test in t4004-match-hwloc.t, ensuring expected
output.