jobspec: add properties support
Problem: RFC14 and RFC25 now supports property-based constraints (via
the "attributes.system.constraints.properties" key) but libjobspec does
not recognize this key.

Add fields to deserialize constraints.properties into.

Throw an exception if a property string contains an invalid character
(! & ' " ^ ` | ( )) or if the value of properties is not a sequence
except when '^' appears as the first character of the property.