writer: add property support
Problem: When a resource vertex (i.e., compute node vertex) with a
property constraint is selected, it must be exported to the resulting
R to enable hierarchical scheduling.

Modify rlite, rv1_nosched, rv1 and JGF writer codes to export
property contraints. Because the resource object written by these
writers can be used to populate a resource graph, it is critical to
export the properties for those writers.