testsuite: move flux-tree to the test suite
Problem: the flux-tree tool has an ambiguous name, has no
documentation, is not user friendly (requires jobspec),
performs a function that is not generally useful, and is not
actively maintained.  As such, it is more of an attractive
nuisance than a benefit to users at this point.

Since the two sharness tests that ostensibly test flux-tree are
probably covering other things in fluxion, leave the tests in place
but relocate flux-tree and its helper to t/scripts and skip the
installation.

If we develop a need for this tool at some point, we should come
up with a new name, find a maintainer, and move it to its own
flux-framework project since it's not fluxion specific.