testsuite: don't save/restore $HOME
Problem: Some tests subvert sharness setting $HOME to
the trash directory, leaving it pointing to the original
$HOME of the invoking user.

The reason given is "sharness modifies $HOME environment variable,
but this interferes with python's package search path, in particular
its user site package".

This would seem to invite the invoking user's personal environment
to interfere with test results, in lieu of providing python
dependencies properly.

Drop the code to restore the user's $HOME.