store: add resource graph duration and setter to metadata
Problem: updated RFC 14 specifies that if the jobspec does
not indicate a duration then the acquired resource expiration
should be propagated to the job time limit. Currently there
is no way to track the current resource expiration in the
resource graph.

Add a struct based on `std::chrono::time_point`s for the
graph start and end times with default values for both.
Default values are defined in `system_defaults.hpp`.
Add a setter to set the times at resource acquisition.