traverser: add checks for duration and handle invalid values
Problem: updated RFC 14 specifies that if a jobspec duration is not set
the resource graph expiration should be propagated to the job's
time limit.

Add validity checks to ensure that if the jobspec duration is longer
than the graph duration the job is not scheduled. Add check to
ensure that the duration (uint64_t) is less than the expressable
int64_t max () value. Check if the scheduled `at` time is negative
(invalid) or greater than or equal to the resource graph end time
(invalid). However, if the job start time plus the duration is
greater than the graph end time we shorten the duration to
fit within the remaining time. We know this is a valid schedule
since the scheduling traversal has returned successfully for the
full duration.