testsuite: add tests for fluxion misconfig
Problem: no tests show that fluxion raises a fatal exception
when the RFC 33 config is not matched with the fluxion config.

Add coverage for
- job submitted with queue when fluxion has no queues
- job submitted with no queue when fluxion has queues

The RFC 33 config would normally block these from reaching the
scheduler, but if the frobnicator is disabled right now we can
get them through.  Remove these tests once fluxion migrates to the
RFC 33 config.