testsuite: fix/simplify annotation test functions
Problem: the t1014-annotations.t validate_sched_annotation()
internal function succeeds when called with any value in the
second argument when sched.t_estimate is not present.  In addition,
the function is rather indirect in its methods and hard to follow,
especially with respect to potential failure cases.

Testing for a valid annotation is simpler now that queues are not
being annotated.  Provide simple functions for checking for the
presence or absence of a valid sched.t_estimate annotation, and
call them in the individual tests.

Add some background to the test description to aid future explorers.