config: update types and define production max duration
Problem: the current system defaults specify a
short SYSTEM_MAX_DURATION of 7 days. Propagating
the value to the resource graph will result in
unexpected behavior and unschedulable jobs as
the current time approaches 7 days from the
startup. The `uint64_t` types differ from
the int64_t used by std::chrono.

Update SYSTEM_MAX_DURATION to be a reasonably
large time in the future and change the
types to comply with std::chrono.