qmanager: support RFC33 TOML queue config
Problem: fluxion queues are currently configured in the qmanager
in the sched-fluxion-qmanager table, but a framework-wide
TOML configuration was proposed in RFC 33.

Implement RFC33 queues in the qmanager by transforming the
RFC 33 configuration into the old sched-fluxion-qmanager syntax.

Raise an error if queues are configured in the sched-fluxion
table instead of in the RFC33-compliant way.

Fixes #950.