testsuite: update flux-tree-helper.py for new OutputFormat
Problem: The flux.util.OutputFormat class in flux-core will get an
improved constructor where `headings` becomes an optional keyword
argument, but this will break the PerfOutputFormat class in
t/scripts/flux-tree-helper.py

Try the new-style constructor arguments in PerfOutputFormat. If that
fails, fall back to the old style so that either works for the
flux-tree-helper.py test script.