libjobspec: add RFC 31 constraint class
Problem: Fluxion does not support the full RFC 31 Job Constraint
Specification.

Add a new Flux::Jobspec::Constraint class which handles validation
and application of a constraint object in jobspec, including the
`properties` operator and `and`, `or`, and `not` boolean logic
operators.

The Constraint base class has two derived types to handle the
supported operations, a PropertyConstraint class which compares
the set of properties in a resource, and a ConditionalConstraint
class which contains a vector of other Constraints and applies
the specified boolean operator when the match() method is called.

Later, other derived classes can be added to handle more constraint
operators.