libjobspec: move parse_error definition to its own header
Problem: If a submodule of libjobspec wishes to use the definition
of Flux::Jobspec::parse_error, it must include jobspec.hpp, but the
submodule header file will most likely need to be included in
jobspec.hpp as well, so this creates a circular header dependency.

Since most submodules will only need the parse_error definition,
just move this to a separate header file.