testsuite: drop recovery-none test
Problem: t1008-recovery-none.t expects the job manager to abort the
scheduler if a job fails to re-allocate resources during the hello
handshake, but this behavior will change soon.

Drop this test.  The behavior it is looking for will either be
addressed by a true fix to #991 or the workaround proposed in
flux-framework/flux-core#4894.