resource: improve errors fro jobpsec parsing
Problem: The specific errors returned from feasibility or alloc request
for jobs with jobspec parse errors, including errors job constraints,
are lost before the error is returned in the RPC to the user or job
validator. Therefore the user only gets a generic error isntead of
a specific, more helpful error.

Pass the exception error message back up through run_match() so it
is available in the error response.

Fixes #1002