policies: replace O(n) accumulate operation
Problem: the current design of the multilevel ID match policy makes use
of a vector of factors (`m_multilevel_factors`) to keep track of all
the added factors and accumulates them to generate a prefix for the
overall vertex score. The O(n) operation is repeated for each vertex.
Furthermore, the vector is not cleared between match traversals, so the
operation becomes increasingly expensive.

Replace the vector with a single `int64_t` which contains the sum of all
vertex factors. Instead of popping the current vertex's factor from a
vector, simply subtract its factor stored in `score_factor_t`. The
change accelerates exclusive matching by a factor of about 6.