planner: ensure result in planner_avail_resources_at
During restart, we can end up in a situation where a time before the
planner's base is requested, in this case negative time, so there's no
earlier time to return from get_state. It feels like get_state should
always return *some kind of state*, but it's not clear how to do that
and avoid accidentally pretending there are no resource available. This
avoids asking for that invalid state by checking the precondition in the
planner, if an `at` before the `plan_start` is requested it's treated as
an invalid argument.