readers: remove subgraph feature
Problem: adding elasicity into fluxion requires the addition of a
new feature to remove subgraphs from the resource graph.

This implementation removes nodes from the resource graph metadata
and clears the node of edges. This makes those nodes essentially
unreachable through traversal or look up from the metadata. Actually
deleting nodes from the resource graph would invalidate the
vertex descriptors and resource graph metadata. This is because
the resource graph is using a vecS instead of a listS for the Boost
adjacency list.