docker: transition bionic to jammy
This is an image update/rework for our CI.  Since bionic has fallen out
of LTS standard support, the original plan was to update to Jammy and
move on.  There's more here than that because Jammy doesn't provide a
386 container build anymore.  To fill in that gap, we also have a Debian
bookworm container, which is now the default target.  There is also a
Jammy build in here to keep an ubuntu LTS current container going, but
the Ubuntu side is painful enough that most of the tests are now on
bookworm.

Additionally:
* el7 is dropped
* fedora outside of 34 (forked to el9) and current are dropped
* docker-checks-run script updated from matching PR in core